requirements.txt
django
validate_email
channels
channels_redis
django-import-export